backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.11.7
email-validator==2.2.0
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.0.0
PyMySQL==1.1.1
alembic==1.14.0
